extract document actions to `useDocumentMenuAction` hook
