+++ app/menus/DocumentMenu.tsx
@@ -1,6 +1,5 @@
 import noop from "lodash/noop";
 import { observer } from "mobx-react";
-import { InputIcon, SearchIcon } from "outline-icons";
 import * as React from "react";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
@@ -11,49 +10,14 @@ import Document from "~/models/Document";
 import { DropdownMenu } from "~/components/Menu/DropdownMenu";
 import { OverflowMenuButton } from "~/components/Menu/OverflowMenuButton";
 import Switch from "~/components/Switch";
-import { ActionV2Separator, createActionV2 } from "~/actions";
-import {
-  pinDocument,
-  createTemplateFromDocument,
-  subscribeDocument,
-  unsubscribeDocument,
-  moveDocument,
-  deleteDocument,
-  permanentlyDeleteDocument,
-  downloadDocument,
-  importDocument,
-  starDocument,
-  unstarDocument,
-  duplicateDocument,
-  archiveDocument,
-  openDocumentHistory,
-  openDocumentInsights,
-  publishDocument,
-  unpublishDocument,
-  printDocument,
-  openDocumentComments,
-  createDocumentFromTemplate,
-  createNestedDocument,
-  shareDocument,
-  copyDocument,
-  searchInDocument,
-  leaveDocument,
-  moveTemplate,
-  restoreDocument,
-  restoreDocumentToCollection,
-  editDocument,
-  applyTemplateFactory,
-} from "~/actions/definitions/documents";
 import useActionContext from "~/hooks/useActionContext";
 import useCurrentUser from "~/hooks/useCurrentUser";
 import useMobile from "~/hooks/useMobile";
 import usePolicy from "~/hooks/usePolicy";
 import useRequest from "~/hooks/useRequest";
 import useStores from "~/hooks/useStores";
-import { ActiveDocumentSection } from "~/actions/sections";
-import { useTemplateMenuActions } from "~/hooks/useTemplateMenuActions";
-import { useMenuAction } from "~/hooks/useMenuAction";
 import { MenuSeparator } from "~/components/primitives/components/Menu";
+import { useDocumentMenuAction } from "~/hooks/useDocumentMenuAction";
 
 type Props = {
   /** Document for which the menu is to be shown */
@@ -148,73 +112,12 @@ function DocumentMenu({
     [user, document]
   );
 
-  const templateMenuActions = useTemplateMenuActions({
+  const rootAction = useDocumentMenuAction({
     document,
+    onFindAndReplace,
+    onRename,
     onSelectTemplate,
   });
-
-  const actions = React.useMemo(
-    () => [
-      restoreDocument,
-      restoreDocumentToCollection,
-      starDocument,
-      unstarDocument,
-      subscribeDocument,
-      unsubscribeDocument,
-      createActionV2({
-        name: `${t("Find and replace")}…`,
-        section: ActiveDocumentSection,
-        icon: <SearchIcon />,
-        visible: !!onFindAndReplace && isMobile,
-        perform: () => onFindAndReplace?.(),
-      }),
-      ActionV2Separator,
-      editDocument,
-      createActionV2({
-        name: `${t("Rename")}…`,
-        section: ActiveDocumentSection,
-        icon: <InputIcon />,
-        visible: !!can.update && !user.separateEditMode && !!onRename,
-        perform: () => requestAnimationFrame(() => onRename?.()),
-      }),
-      shareDocument,
-      createNestedDocument,
-      importDocument,
-      createTemplateFromDocument,
-      duplicateDocument,
-      publishDocument,
-      unpublishDocument,
-      archiveDocument,
-      moveDocument,
-      moveTemplate,
-      applyTemplateFactory({ actions: templateMenuActions }),
-      pinDocument,
-      createDocumentFromTemplate,
-      ActionV2Separator,
-      openDocumentComments,
-      openDocumentHistory,
-      openDocumentInsights,
-      downloadDocument,
-      copyDocument,
-      printDocument,
-      searchInDocument,
-      ActionV2Separator,
-      deleteDocument,
-      permanentlyDeleteDocument,
-      leaveDocument,
-    ],
-    [
-      t,
-      isMobile,
-      templateMenuActions,
-      can.update,
-      user.separateEditMode,
-      onFindAndReplace,
-      onRename,
-    ]
-  );
-
-  const rootAction = useMenuAction(actions);
 
   const context = useActionContext({
     isContextMenu: true,
