+++ shared/i18n/locales/en_US/translation.json
@@ -539,6 +539,7 @@
   "Video": "Video",
   "None": "None",
   "Delete embed": "Delete embed",
+  "Rename": "Rename",
   "Could not import file": "Could not import file",
   "Unsubscribed from document": "Unsubscribed from document",
   "Unsubscribed from collection": "Unsubscribed from collection",
@@ -561,7 +562,6 @@
   "A-Z sort": "A-Z sort",
   "Z-A sort": "Z-A sort",
   "Manual sort": "Manual sort",
-  "Rename": "Rename",
   "Collection menu": "Collection menu",
   "Comment options": "Comment options",
   "Enable embeds": "Enable embeds",
